refactor(studio): move 'SongRouteParams' to common types
Centralize the route params type and update all imports to the shared location.
