+++ apps/studio/src/pages/home/library/MarketplaceTab/SoldOutSale.tsx
@@ -5,7 +5,7 @@ import { Alert, Button, HorizontalLine } from "@newm-web/elements";
 import theme from "@newm-web/theme";
 import { Sale } from "@newm-web/types";
 import { EndSaleModal } from "./EndSaleModal";
-import { SongRouteParams } from "../types";
+import { SongRouteParams } from "../../../../common/types";
 import { useEndSaleThunk } from "../../../../modules/sale";
 
 interface SoldOutSaleProps {
