--- apps/studio/src/pages/home/library/types.ts
@@ -1,3 +1,0 @@
-export interface SongRouteParams {
-  songId: string;
-}
