--- apps/studio/src/pages/home/releases/types.ts
@@ -1,3 +1,0 @@
-export interface SongRouteParams {
-  songId: string;
-}
