+++ apps/studio/src/pages/home/uploadSong/AdvancedSongDetails.tsx
@@ -30,7 +30,7 @@ import {
 } from "../../../common";
 import { emptyProfile, useGetProfileQuery } from "../../../modules/session";
 import { CoverRemixSample } from "../../../components";
-import { SongRouteParams } from "../library/types";
+import { SongRouteParams } from "../../../common/types";
 
 const AdvancedSongDetails = () => {
   const { data: { firstName } = emptyProfile } = useGetProfileQuery();
